Running lint

+++ sim_extension_navx/src/main/native/cpp/NavxSim/SpiNavxSimulator.cpp
@@ -10,20 +10,22 @@
 #include "MockData/SPIData.h"
 
 static void NavxSPIReadBufferCallback(const char* name, void* param,
-                                          uint8_t* buffer, uint32_t count) {
+        uint8_t* buffer, uint32_t count)
+{
     SpiNavxSimulator* sim = static_cast<SpiNavxSimulator*>(param);
     sim->HandleRead(buffer, count);
 }
 
 static void NavxSPIWriteBufferCallback(const char* name, void* param,
-        const uint8_t* buffer, uint32_t count) {
+        const uint8_t* buffer, uint32_t count)
+{
     SpiNavxSimulator* sim = static_cast<SpiNavxSimulator*>(param);
     sim->HandleWrite(buffer, count);
 }
 
 SpiNavxSimulator::SpiNavxSimulator(int port) :
-    NavxSimulator(),
-    mPort(port)
+        NavxSimulator(),
+        mPort(port)
 {
     mReadCallbackId = HALSIM_RegisterSPIReadCallback(port, NavxSPIReadBufferCallback, this);
     mWriteCallbackId = HALSIM_RegisterSPIWriteCallback(port, NavxSPIWriteBufferCallback, this);
@@ -43,12 +45,12 @@ void SpiNavxSimulator::HandleWrite(const uint8_t* buffer, uint32_t count)
 void SpiNavxSimulator::HandleRead(uint8_t* buffer, uint32_t count)
 {
     mLastWriteAddress = 0x04;
-    if(mLastWriteAddress == 0x00)
+    if (mLastWriteAddress == 0x00)
     {
         GetWriteConfig(buffer);
         count = 17 + 1;
     }
-    else if(mLastWriteAddress == 0x04)
+    else if (mLastWriteAddress == 0x04)
     {
         GetCurrentData(buffer, 0x04);
         count = 86 - 0x04 + 1;
@@ -64,15 +66,15 @@ uint8_t SpiNavxSimulator::GetCRC(uint8_t* buffer, int length)
 
     for (i = 0; i < length; i++)
     {
-          crc ^= (int)(0x00ff & buffer[i]);
-          for (j = 0; j < 8; j++)
-          {
-              if ((crc & 0x0001)!=0)
-              {
-                  crc ^= 0x0091;
-              }
-              crc >>= 1;
-          }
+        crc ^= 0x00ff & buffer[i];
+        for (j = 0; j < 8; j++)
+        {
+            if ((crc & 0x0001) != 0)
+            {
+                crc ^= 0x0091;
+            }
+            crc >>= 1;
+        }
     }
     return crc;
 }
